Move ODSolution to its own submodule
