+++ src/od/process/mod.rs
@@ -40,7 +40,6 @@ pub use trigger::EkfTrigger;
 mod rejectcrit;
 use self::msr::TrackingDataArc;
 pub use self::rejectcrit::ResidRejectCrit;
-mod export;
 mod solution;
 pub use solution::ODSolution;
